Found Deauth script hehe
Turns out I had the script that I previously thought was lost forever in a backup drive.  Originally I intended for this project to kick off a specific device from WiFi but due to a lack of documentation related to the MAC ranges of  devices, I had to scale it down to targeting a specific guest.